refactor to put websocket code in a separate subdirectory
